+++ js/constants.js
@@ -1,0 +1,29 @@
+const PHOTOS_COUNT = 25;
+const LIKES_MIN_COUNT = 15;
+const LIKES_MAX_COUNT = 200;
+const COMMENTS_COUNT = 15;
+const COMMENTS_TEXT = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Отличный кадр!', 'Меня там точно не хватает!'
+];
+const DESCRIPTIONS = [
+  'На память!', 'Рандом фото вам в ленту )', 'Красота-то какая ! Ляпота!', 'И такое бывает.', 'Отдыхать, так на полную!',
+  'Хорошо там где нас нет.', 'Вполне не плохо!', 'Придумайте подпись к фото )'
+];
+const NAMES = [
+  'Николай','Иван','Харитон','Ульяна','Яков','Нина','Егор','Педро','Оля','Никитос','Ягодка','Тихон','Николь','Олег'
+];
+const AVATAR_COUNT = 6;
+
+export {PHOTOS_COUNT};
+export {LIKES_MIN_COUNT};
+export {LIKES_MAX_COUNT};
+export {COMMENTS_COUNT};
+export {COMMENTS_TEXT};
+export {DESCRIPTIONS};
+export {NAMES};
+export {AVATAR_COUNT};
